FaceRDjango/requirements.txt
asgiref==3.2.3
Click==7.0
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
dlib==19.19.0
face-recognition==1.2.3
face-recognition-models==0.3.0
jsonfield==2.0.2
numpy==1.18.1
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
